Minor changes, still not in a workable state.
The instance of Question held by Game at any given time seems to
suddenly be able to reference a null value without warning. This is
likely a simple matter, which I'll look into in the morning.
QuestionsLoader is not yet workable, so for now Game still just
attempts to load the included xml file with questions. Code from the
loader static method has been modified, and now seems to work out
better than before, using a safer temporary incrementation of
currentAnswerNumber as a default value when fetching the tag's ID.
Button drawables have been changed, the previous ones did not display
correctly on the device, unlike in the emulator or in Eclipse (for an
as-yet unknown reason). Could be Samsung Galaxy S2-specific though.
Made changes to allow automatic adaptation to a change in orientation,
although I may have to handle orientation myself, if things start to go
wrong when changing orientation.

In general, this is still a non-working commit of the work done so far.
Hopefully this goes somewhere...